add decimal support for ratings

+++ cmd/migrate/migrations/20241124223305_add-product-table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS products (
   `title` VARCHAR(255) NOT NULL,
   `description` TEXT NOT NULL,
   `seller` VARCHAR(255) NOT NULL,
-  `rating` TINYINT NOT NULL CHECK (rating >= 1 AND rating <= 5),
+  `rating` DECIMAL(2, 1) NOT NULL CHECK (rating >= 1.0 AND rating <= 5.0),
   `createdAt` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (id)
 );
